+++ universal-login-example/src/App.tsx
@@ -6,9 +6,7 @@ const config = {
   relayerUrl: process.env.RELAYER_URL!,
   jsonRpcURl: process.env.JSON_RPC_URL!,
   ensDomains: [process.env.ENS_DOMAIN_1!],
-}
-
-console.log(config);
+};
 
 const universalLogin = new ULWeb3Provider(
   new Web3.providers.HttpProvider(config.jsonRpcURl),
@@ -17,10 +15,6 @@ const universalLogin = new ULWeb3Provider(
 );
 const web3 = new Web3(universalLogin);
 
-web3.eth.getAccounts()
-  .then(accounts => console.log('accounts', accounts))
-  .catch(console.error);
-
 function App() {
   function sendTx() {
     web3.eth.sendTransaction({
@@ -28,8 +22,7 @@ function App() {
       to: '0x7ffC57839B00206D1ad20c69A1981b489f772031',
       value: '500000000000000',
     })
-      .then(console.log)
-      .catch(console.error);
+      .then(console.log, console.error);
   }
 
   function create() {
