+++ universal-login-web3/lib/utils.ts
@@ -1,0 +1,12 @@
+import {Property} from 'reactive-properties';
+
+export function waitFor<T>(predicate: (value: T) => boolean): (prop: Property<T>) => Promise<void> {
+  return (source) => new Promise((resolve) => {
+    const unsubscribe = source.subscribe(() => {
+      if (predicate(source.get())) {
+        resolve();
+        unsubscribe();
+      }
+    });
+  });
+}
